LayoutTests/imported/w3c/web-platform-tests/reporting/cross-origin-reports-isolated.https.sub-expected.txt



FAIL Reports were sent in two requests. assert_equals: expected 2 but got 0
